restructure tests to reflect new structure, create test_converters.py for converter-specific tests
