+++ tests/test_indicators.py
@@ -537,7 +537,7 @@ def test_signature():
     assert sig.parameters["thresh"].kind == sig.parameters["thresh"].KEYWORD_ONLY
     assert sig.return_annotation == xr.DataArray
 
-    sig = signature(xclim.atmos.wind_speed_from_vector)
+    sig = signature(xclim.convert.wind_speed_from_vector)
     assert sig.return_annotation == tuple[xr.DataArray, xr.DataArray]
 
 
@@ -585,7 +585,7 @@ def test_parse_doc():
     assert "references" not in doc
     assert doc["long_name"] == "The mean daily temperature at the given time frequency."
 
-    doc = parse_doc(xclim.indices.saturation_vapor_pressure.__doc__)
+    doc = parse_doc(xclim.indices.converters.saturation_vapor_pressure.__doc__)
     assert doc["parameters"]["ice_thresh"]["description"] == (
         "Threshold temperature under which to switch to equations in reference to ice instead of water. "
         "If None (default) everything is computed with reference to water. "
