+++ tests/test_temperature.py
@@ -8,7 +8,7 @@ import xarray as xr
 from packaging.version import Version
 from pint import __version__ as __pint_version__
 
-from xclim import atmos
+from xclim import atmos, convert
 from xclim.core.calendar import percentile_doy
 from xclim.core.options import set_options
 from xclim.core.units import convert_units_to
@@ -168,8 +168,8 @@ class TestTmean:
     def test_Tmean_3d_data(self, open_dataset):
         ds_tmax = open_dataset(self.nc_files[0])
         ds_tmin = open_dataset(self.nc_files[1])
-        tas = atmos.tg(ds_tmin.tasmin, ds_tmax.tasmax)
-        tas_C = atmos.tg(ds_tmin.tasmin, ds_tmax.tasmax)
+        tas = convert.tg(ds_tmin.tasmin, ds_tmax.tasmax)
+        tas_C = convert.tg(ds_tmin.tasmin, ds_tmax.tasmax)
         tas_C.values -= K2C
         tas_C.attrs["units"] = "C"
         # put a nan somewhere
@@ -1398,8 +1398,8 @@ def test_corn_heat_units(open_dataset):
         tnC = tn - K2C
         tnC.attrs["units"] = "C"
 
-    chu = atmos.corn_heat_units(tasmin=tn, tasmax=tx, thresh_tasmin="4.44 degC", thresh_tasmax="10 degC")
-    chuC = atmos.corn_heat_units(tasmin=tnC, tasmax=tx, thresh_tasmin="4.44 degC", thresh_tasmax="10 degC")
+    chu = convert.corn_heat_units(tasmin=tn, tasmax=tx, thresh_tasmin="4.44 degC", thresh_tasmax="10 degC")
+    chuC = convert.corn_heat_units(tasmin=tnC, tasmax=tx, thresh_tasmin="4.44 degC", thresh_tasmax="10 degC")
 
     np.testing.assert_allclose(chu, chuC, rtol=1e-3)
