+++ tests/test_wind.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import numpy as np
 import pytest
 
-from xclim import atmos
+from xclim import atmos, convert
 
 
 class TestWindSpeedIndicators:
@@ -11,7 +11,7 @@ class TestWindSpeedIndicators:
 
     def test_calm_windy_days(self, open_dataset):
         with open_dataset(self.test_data) as ds:
-            sfcwind, _ = atmos.wind_speed_from_vector(ds.uas, ds.vas, calm_wind_thresh="0 m/s")
+            sfcwind, _ = convert.wind_speed_from_vector(ds.uas, ds.vas, calm_wind_thresh="0 m/s")
             calm = atmos.calm_days(sfcwind, thresh="5 m/s")
             windy = atmos.windy_days(sfcwind, thresh="5 m/s")
             c = sfcwind.resample(time="MS").count()
@@ -27,7 +27,7 @@ class TestSfcWind:
     )
     def test_sfcWind(self, open_dataset, metric):
         with open_dataset(self.test_data) as ds:
-            sfcWind, _ = atmos.wind_speed_from_vector(ds.uas, ds.vas)
+            sfcWind, _ = convert.wind_speed_from_vector(ds.uas, ds.vas)
             sfcWind_calculated = getattr(atmos, f"sfcWind_{metric}")(sfcWind)
 
             resample = sfcWind.resample(time="YS")
@@ -44,7 +44,7 @@ class TestSfcWindMax:
     )
     def test_sfcWindmax(self, open_dataset, metric):
         with open_dataset(self.test_data) as ds:
-            sfcWind, _ = atmos.wind_speed_from_vector(ds.uas, ds.vas)
+            sfcWind, _ = convert.wind_speed_from_vector(ds.uas, ds.vas)
             sfcWindmax_calculated = getattr(atmos, f"sfcWindmax_{metric}")(sfcWind)
 
             resample = sfcWind.resample(time="YS")
